miagrey

+++ app.js
@@ -93,10 +93,11 @@ bot.on("message", async message => {
   if(message.author.bot){
     if(message.member.roles){
       if(message.member.roles.some(r=>["Mantaro"].includes(r.name))){
-        console.log("channel: " + message.channel.name);
-        message.delete()
-        .then(msg => console.log(`Deleted message from ${msg.author.username}`))
-        .catch(console.error);
+        if(message.channel.name == "📵канализация"){
+          message.delete()
+          .then(msg => console.log(`Удалено сообщение от ${msg.author.username}`))
+          .catch(console.error);
+        }
       }
     }
     return;
